pyproject.toml
[project]
name = "betrayal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cryptography>=46.0.6",
    "doit>=0.37.0",
    "faiss-cpu>=1.13.2",
    "jinja2>=3.1.6",
    "jsonschema>=4.26.0",
    "openai>=2.30.0",
    "python-dotenv>=1.2.2",
    "pytest>=8.3.5",
    "tenacity>=9.1.4",
    "tiktoken>=0.12.0",
]

[tool.pytest.ini_options]
python_files = ["test_*.py"]
testpaths = ["tests"]
